backend/requirements.txt
Django==2.2.10
django-cors-headers==2.5.2
django-extensions==2.2.8
django-model-utils==4.0.0
djangorestframework==3.9.2
coverage==5.0.3
flake8==3.7.9
psycopg2==2.8.5
honcho==1.0.1
ipython==7.14.0
pillow==7.1.2